move tests to the helper test class
